Moved vsg::NemorySlots out into it's own header/source file. Added logging of MemorySlots actions for debugging purposes. Added extra debug output support for tracking memory allocation/deallocations.
