+++ src/vsg/core/Allocator.cpp
@@ -19,272 +19,15 @@ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLI
 
 using namespace vsg;
 
-#if 0
+#define DO_CHECK 0
+
+#if DO_CHECK
 #    define DEBUG \
         if (true) std::cout
 #else
 #    define DEBUG \
         if (false) std::cout
 #endif
-
-///////////////////////////////////////////////////////////////////////////////
-//
-// MemorySlots
-//
-MemorySlots::MemorySlots(size_t availableMemorySize)
-{
-    _availableMemory.insert(SizeOffset(availableMemorySize, 0));
-    _offsetSizes.insert(OffsetSize(0, availableMemorySize));
-
-    _totalMemorySize = availableMemorySize;
-}
-
-size_t MemorySlots::totalAvailableSize() const
-{
-    size_t totalSize = 0;
-    for (const auto& sizeOffset : _availableMemory)
-    {
-        totalSize += sizeOffset.first;
-    }
-    return totalSize;
-}
-
-size_t MemorySlots::totalReservedSize() const
-{
-    size_t totalSize = 0;
-    for (const auto& sizeOffset : _reservedOffsetSizes)
-    {
-        totalSize += sizeOffset.second;
-    }
-    return totalSize;
-}
-
-bool MemorySlots::check() const
-{
-    if (_availableMemory.size() != _offsetSizes.size())
-    {
-        std::cout << "Warning: MemorySlots::check() _availableMemory.size() " << _availableMemory.size() << " != _offsetSizes.size() " << _offsetSizes.size() << std::endl;
-    }
-
-    size_t availableSize = 0;
-    for (auto& offsetSize : _offsetSizes)
-    {
-        availableSize += offsetSize.second;
-    }
-
-    size_t reservedSize = 0;
-    for (auto& offsetSize : _reservedOffsetSizes)
-    {
-        reservedSize += offsetSize.second;
-    }
-
-    size_t computedSize = availableSize + reservedSize;
-    if (computedSize != _totalMemorySize)
-    {
-        std::cout << "Warning : MemorySlots::check() " << this << " failed, computeedSize (" << computedSize << ") != _totalMemorySize (" << _totalMemorySize << ")" << std::endl;
-
-        report();
-
-        throw "MemorySlots check failed";
-
-        return false;
-    }
-
-    return true;
-}
-
-void MemorySlots::report() const
-{
-    std::cout << "MemorySlots::report()" << std::endl;
-    for (auto& [offset, size] : _offsetSizes)
-    {
-        std::cout << "    available " << offset << ", " << size << std::endl;
-    }
-
-    for (auto& [offset, size] : _reservedOffsetSizes)
-    {
-        std::cout << "    reserved " << std::dec << offset << ", " << size << std::endl;
-    }
-}
-
-MemorySlots::OptionalOffset MemorySlots::reserve(size_t size, size_t alignment)
-{
-    if (full()) return OptionalOffset(false, 0);
-
-    auto itr = _availableMemory.lower_bound(size);
-    while (itr != _availableMemory.end())
-    {
-        SizeOffset slot(*itr);
-        size_t slotStart = slot.second;
-        size_t slotSize = slot.first;
-        size_t slotEnd = slotStart + slotSize;
-
-        size_t alignedStart = ((slotStart + alignment - 1) / alignment) * alignment;
-        size_t alignedEnd = alignedStart + size;
-        size_t alignedSize = alignedEnd - slotStart;
-
-        if (alignedSize <= slotSize)
-        {
-            // remove slot
-            _availableMemory.erase(itr);
-            if (auto offsetSize_itr = _offsetSizes.find(slotStart); offsetSize_itr != _offsetSizes.end()) _offsetSizes.erase(offsetSize_itr);
-
-            // check if there the front of the slot isn't used completely, if so generate an available space for it.
-            if (alignedStart > slotStart)
-            {
-                // insert new slot with previous slots start and new end.
-                size_t preAlignedStartSize = alignedStart - slotStart;
-                _availableMemory.insert(SizeOffset(preAlignedStartSize, slotStart));
-                _offsetSizes.insert(OffsetSize(slotStart, preAlignedStartSize));
-            }
-
-            // check if there is space at the end slot that isn't used completely, if so generate an available space for it.
-            if (alignedEnd < slotEnd)
-            {
-                // insert new slot with new end and new size
-                size_t postAlignedEndSize = slotEnd - alignedEnd;
-                _availableMemory.insert(SizeOffset(postAlignedEndSize, alignedEnd));
-                _offsetSizes.insert(OffsetSize(alignedEnd, postAlignedEndSize));
-            }
-
-            _reservedOffsetSizes[alignedStart] = (alignedEnd - alignedStart);
-
-#if DO_CHECK
-            check();
-#endif
-
-            return OptionalOffset(true, alignedStart);
-        }
-        else
-        {
-            // std::cout << "    Slot slotStart = " << slotStart << ", slotSize = " << slotSize << " not big enough once for request size = " << size << std::endl;
-            ++itr;
-        }
-    }
-
-    //std::cout<<"MemorySlots::reserve("<<std::dec<<size<<") with alingment "<<alignment<<" No slots available for this size, biggest available slot is : "<<_availableMemory.rbegin()->first<<std::endl;
-    //report();
-
-    return OptionalOffset(false, 0);
-}
-
-void MemorySlots::release(size_t offset, size_t size)
-{
-    auto reserved_itr = _reservedOffsetSizes.find(offset);
-    if (reserved_itr == _reservedOffsetSizes.end())
-    {
-#if DO_CHECK
-        std::cout << "   MemorySlots::release() slot not found" << std::endl;
-#endif
-        return;
-    }
-    else
-    {
-#if DO_CHECK
-        if (reserved_itr->second != size)
-        {
-            std::cout << "    MemorySlots::release() slot found but sizes are inconsistent reserved_itr->second = " << std::dec << reserved_itr->second << ", size=" << size << std::endl;
-            if (size != 0) throw "MemorySlots::release() slot found but sizes are inconsistent reserved_itr->second";
-        }
-        else
-        {
-            std::cout << "    MemorySlots::release() slot found " << std::endl;
-        }
-#endif
-
-        size = reserved_itr->second;
-
-        _reservedOffsetSizes.erase(reserved_itr);
-    }
-
-    if (_offsetSizes.empty())
-    {
-        // first empty space
-        _availableMemory.insert(SizeOffset(size, offset));
-        _offsetSizes.insert(OffsetSize(offset, size));
-
-#if DO_CHECK
-        check();
-#endif
-        return;
-    }
-
-    // need to find adjacent blocks before and after to see if we can join them together options are:
-    //    abutes to neither before or after
-    //    abutes to before, so replace before with new combined length
-    //    abutes to after, so remove after entry and insert new entry with combined length
-    //    abutes to both before and after, so replace before with newly combined length of all three, remove after entry
-
-    auto slotAfter = _offsetSizes.upper_bound(offset);
-
-    auto slotBefore = slotAfter;
-    if (slotBefore != _offsetSizes.end())
-    {
-        if (slotBefore == _offsetSizes.begin())
-        {
-            slotBefore = _offsetSizes.end();
-        }
-        else
-        {
-            --slotBefore;
-        }
-    }
-    else
-    {
-        slotBefore = _offsetSizes.rbegin().base();
-    }
-
-    auto eraseSlot = [&](OffsetSizes::iterator offsetSizeItr) {
-        auto range = _availableMemory.equal_range(offsetSizeItr->second);
-        for (auto itr = range.first; itr != range.second; ++itr)
-        {
-            if (itr->second == offsetSizeItr->first)
-            {
-                _availableMemory.erase(itr);
-                _offsetSizes.erase(offsetSizeItr);
-                break;
-            }
-        }
-    };
-
-    if (slotBefore != _offsetSizes.end())
-    {
-        size_t endOfBeforeSlot = slotBefore->first + slotBefore->second;
-
-        if (endOfBeforeSlot == offset)
-        {
-            size_t endOfReleasedSlot = offset + size;
-            size_t totalSizeOfMergedSlots = endOfReleasedSlot - slotBefore->first;
-
-            offset = slotBefore->first;
-            size = totalSizeOfMergedSlots;
-
-            eraseSlot(slotBefore);
-        }
-    }
-    if (slotAfter != _offsetSizes.end())
-    {
-        size_t endOfReleasedSlot = offset + size;
-
-        if (endOfReleasedSlot == slotAfter->first)
-        {
-            size_t endOfSlotAfter = slotAfter->first + slotAfter->second;
-            size_t totalSizeOfMergedSlots = endOfSlotAfter - offset;
-            size = totalSizeOfMergedSlots;
-
-            eraseSlot(slotAfter);
-        }
-    }
-
-    _availableMemory.insert(SizeOffset(size, offset));
-    _offsetSizes.insert(OffsetSize(offset, size));
-
-    //report();
-
-#if DO_CHECK
-    check();
-#endif
-}
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////
 //
@@ -298,9 +41,9 @@ Allocator::Allocator(std::unique_ptr<Allocator> in_nestedAllocator) :
     allocatorMemoryBlocks.resize(vsg::ALLOCATOR_LAST);
 
     // TODO need to set to a more sensible default
-    allocatorMemoryBlocks[vsg::ALLOCATOR_OBJECTS].reset(new MemoryBlocks("ALLOCATOR_OBJECTS", size_t(1024 * 1024 * 16)));
-    allocatorMemoryBlocks[vsg::ALLOCATOR_DATA].reset(new MemoryBlocks("ALLOCATOR_DATA", size_t(1024 * 1024 * 16)));
-    allocatorMemoryBlocks[vsg::ALLOCATOR_NODES].reset(new MemoryBlocks("ALLOCATOR_NODES", size_t(1024 * 1024 * 16)));
+    allocatorMemoryBlocks[vsg::ALLOCATOR_OBJECTS].reset(new MemoryBlocks("ALLOCATOR_OBJECTS", size_t(4096)));
+    allocatorMemoryBlocks[vsg::ALLOCATOR_DATA].reset(new MemoryBlocks("ALLOCATOR_DATA", size_t(2048)));
+    allocatorMemoryBlocks[vsg::ALLOCATOR_NODES].reset(new MemoryBlocks("ALLOCATOR_NODES", size_t(512)));
 }
 
 Allocator::~Allocator()
@@ -314,21 +57,21 @@ std::unique_ptr<Allocator>& Allocator::instance()
     return s_allocator;
 }
 
-void Allocator::report() const
+void Allocator::report(std::ostream& out) const
 {
-    DEBUG << "Allocator::report() " << allocatorMemoryBlocks.size() << std::endl;
+    out << "Allocator::report() " << allocatorMemoryBlocks.size() << std::endl;
     std::scoped_lock<std::mutex> lock(mutex);
     for (const auto& memoryBlocks : allocatorMemoryBlocks)
     {
         if (memoryBlocks)
         {
-            DEBUG << "    " << memoryBlocks->name;
+            out << "    " << memoryBlocks->name;
             for (const auto& memoryBlock : memoryBlocks->memoryBlocks)
             {
                 const auto& memorySlots = memoryBlock->memorySlots;
-                DEBUG << ", [" << memorySlots.totalReservedSize() << ", " << memorySlots.maximumAvailableSpace() << "]";
+                out << ", [" << memorySlots.totalReservedSize() << ", " << memorySlots.maximumAvailableSpace() << "]";
             }
-            DEBUG << std::endl;
+            out << std::endl;
         }
     }
 }
@@ -409,7 +152,10 @@ bool Allocator::MemoryBlock::deallocate(void* ptr)
         size_t offset = static_cast<uint8_t*>(ptr) - memory;
         if (offset < memorySlots.totalMemorySize())
         {
-            memorySlots.release(offset, 0);
+            if (!memorySlots.release(offset, 0))
+            {
+                std::cout<<"Allocator::MemoryBlock::deallocate("<<ptr<<") problem - couldn't release"<<std::endl;
+            }
             return true;
         }
     }
